requirements.txt
termcolor==2.4.0
PyYAML==6.0.1
requests==2.31.0
beautifulsoup4==4.12.3